Modify home view controller
