--- SwiftTemplate/Modules/Home/View/News/CollectionTableViewCell.swift
@@ -1,122 +1,0 @@
-//
-//  NewsTableViewCell.swift
-//  SwiftTemplate
-//
-//  Created by Cristian Sancricca on 31/05/2022.
-//
-
-import UIKit
-
-class CollectionTableViewCell: UITableViewCell {
-    
-    //MARK: - Properties
-    
-    static let idenfifier = "NewsTableViewCell"
-    
-    private var timer : Timer?
-    private var currentCellIndex = 0
-    
-    lazy var viewModel: HomeViewModel = {
-        let homeViewModel = HomeViewModel()
-        homeViewModel.delegate = self
-        return homeViewModel
-    }()
-    
-    private let collectionView: UICollectionView = {
-        let layout = UICollectionViewFlowLayout()
-        layout.scrollDirection = .horizontal
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collectionView.register(NewsCollectionViewCell.self, forCellWithReuseIdentifier: NewsCollectionViewCell.identifier)
-        collectionView.showsHorizontalScrollIndicator = false
-        return collectionView
-    }()
-    
-    //MARK: - Lifecycle
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        contentView.backgroundColor = .red
-        contentView.addSubview(collectionView)
-        
-        viewModel.getNewsData()
-        
-        collectionView.delegate = self
-        collectionView.dataSource = self
-        
-        startTimer()
-    }
-    
-    func startTimer(){
-        timer = Timer.scheduledTimer(timeInterval: 2.5, target: self, selector: #selector(moveNextIndex), userInfo: nil, repeats: true)
-    }
-    
-    @objc func moveNextIndex(){
-        if currentCellIndex < 4 {
-            currentCellIndex += 1
-        }else {
-            currentCellIndex = 0
-        }
-        collectionView.scrollToItem(at: IndexPath(item: currentCellIndex, section: 0), at: .centeredHorizontally, animated: true)
-    }
-    
-    override func layoutSubviews() {
-        super.layoutSubviews()
-        collectionView.frame = contentView.bounds
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        
-    }
-
-    override func setSelected(_ selected: Bool, animated: Bool) {
-        super.setSelected(selected, animated: animated)
-
-    }
-
-}
-
-//MARK: - CollectionView Delegate & DataSource
-
-extension CollectionTableViewCell: UICollectionViewDelegate, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout{
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: collectionView.frame.width, height: 250)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: NewsCollectionViewCell.identifier, for: indexPath) as? NewsCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-        let new = viewModel.getNews(at: indexPath.row)
-        cell.configureCell(with: new)
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return viewModel.getNewsCount()
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, minimumLineSpacingForSectionAt section: Int) -> CGFloat {
-        return 0
-    }
-    
-}
-
-extension CollectionTableViewCell: HomeViewModelDelegate {
-    
-    func didGetNewsData() {
-        DispatchQueue.main.async { [weak self] in
-            self?.collectionView.reloadData()
-        }
-    }
-    
-    func didFailGettingNewsData(error: String) {
-        
-    }
-}
